backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0

